[Feat] #57 - CongreteModal, EmitterView 파일 분리
